Network Analysis Algorithms Overview

There are a few dynamical properties of particular interest that network approaches can shed some light on.

Existence + uniqueness of steady states

Persistence

Concentration Robustness

Algorithms used by this package to decide these properties are detailed on this page. A significant number of the results can be found in the Foundations of Chemical Reaction Network Theory textbook, by Martin Feinberg[1].

In general these properties will depend on the choices of parameters for the system and on the initial conditions (which can be provided as a dictionary, vector, or tuple of Pair, as described in the homepage). Many networks have multiple steady states for some set of rate constants but not others.

Additionally, the dynamical properties of the system may depend on the initial conditions of the network, which define the stoichiometric compatibility class that the network evolves in. The stoichiometric compatibility class is the set of points in concentration space that are reachable from a given initial condition, which is an affine subspace $u_0 + S$

When we say that a network is multistationary, or has the capacity for multiple steady states, what that means is that there is some stoichiometric compatibility class for which this is true. This means that there may be only one or zero steady states in some other stoichiometric compatibility class. The SCC can be inferred from the initial concentrations of the species in a reaction network via the network's conservation laws.

The summary function for the package is networksummary.

CatalystNetworkAnalysis.networksummary — Function

networksummary(rn::ReactionSystem; p::VarMapType, u0::VarMapType, mv = false)

Summary of properties that can be inferred from the structure of a reaction network. May give different results depending on whether `p` or `u0` is supplied. The set of functions run are the following: 
- `hasuniquesteadystates`
- `isconcentrationrobust`
- `ispersistent`
- `mixedvolume`

Mixed volume may take a very long time to run and is disabled by default. It can be enabled by setting the flag `mv = true`. Note that mixed volume requires an initial condition.

Uniqueness of Steady States

The functionality that decides uniqueness of steady states is in the function hasuniqueequilibria(rn). The uniqueness of steady states checks proceeds as follows. The function terminates every time a network is conclusively determined to have unique steady states or multi-stationarity.

Deficiency one and deficiency zero theorems: a reaction network that satisfies the conditions of these two theorems will be guaranteed to have a unique steady state for every stoichiometric compatibility class.[1]

Deficiency one and higher deficiency algorithm: Both of these algorithms construct inequality systems that, if satsifiable, would indicate that a reaction network has the ability to admit multiple steady states in a given stoichiometric compatibility class.[2]

Concordance: This is a graph property that guarantees the uniqueness of steady states, since it guarantees that the species-formation rate function is injective. As a result, there is only one solution to $\frac{dc}{dt} = 0$.[2]

Future desiderata for this check include methods from monotone systems theory,degree theory, and embedded multi-stationary networks - see the [Roadmap] for the package.

Other functionality related to the existence and uniqueness of steady states:

CatalystNetworkAnalysis.hasuniquesteadystates — Function

hasuniquesteadystates(rn::ReactionSystem)

Check whether a reaction network has the capacity to admit multiple steady states, for some choice of rate constants. Return codes: 
- :NO_EQUILIBRIUM - no positive equilibrium for any choice of rate constants
- :STRUCTURALLY_UNIQUE - only one steady state for every SCC, for every choice of rate constants
- :STRUCTURALLY_MULTIPLE - multiple steady states in a certain SCC guaranteed for any choice of rate constants
- :KINETICALLY_MULTIPLE - multiple steady states in a certain SCC guaranteed for certain choices of rate constants
- :POSSIBLY_MULTIPLE - discordant and/or high deficiency, but inconclusive whether there are system parameters that lead to the existence of an SCC with multiple steady states.

CatalystNetworkAnalysis.mixedvolume — Function

mixedvolume(rn::ReactionSystem, u0::VarMapType)

Compounds an upper bound on the number of steady states in a particular stoichiometric compatibility class.

CatalystNetworkAnalysis.isconcordant — Function

isconcordant(rn::ReactionSystem, atol=1e-12)

Given a reaction network (and an absolute tolerance for the nullspace matrix below which entries should be zero), test whether the reaction network's graph has a property called concordance. A concordant network will not admit multiple equilibria in any stoichiometric compatibility class. The algorithm for this check follows Haixia Ji's PhD thesis, (Ji, 2011).

CatalystNetworkAnalysis.deficiencyonealgorithm — Function

deficiencyonealgorithm(rn::ReactionSystem)

Determine whether a regular deficiency one network will have the ability to admit multiple equilibria and degenerate equilibria. Returns true if so.

Additionally, one might check the functions iscomplexbalanced() and isdetailedbalanced() implemented in the main Catalyst package, which work for kinetic systems (reaction networks with an assignment of parameter values). Both of these conditions are sufficient to guarantee that the reaction network will have unique steady states, for a given choice of parameters, and that these steady states will be asymptotically stable within each stoichiometric compatibility class.[1]

Stability

Bistability is a very important property of reaction networks, since switchlike behavior in biological networks is realized through bistability. The network analysis package currently does not have any tooling for stability analysis, but please see this section of the main Catalyst documentation for some tips.

Persistence

A chemical reaction network is persistent if each species does not "die out" - that is, if the initial condition of the network has positive concentration in a given species, that species' concentration will never go to zero.

It is not currently possible to conclusively determine persistence in every case. The approach used by this package checks the necessary conditions and sufficient conditions outlined in this paper.[4] Many persistent networks will be detectably persistent via this algorithm, and many non-persistent networks will be detectably non-persistent, but the function will return inconclusive in many cases.

CatalystNetworkAnalysis.ispersistent — Function

ispersistent(rs::ReactionSystem)

Checks if a reaction system is persistent, meaning that none of its species with positive concentration will go extinct (asymptotically approach 0). The possible outputs: 
- :PERSISTENT
- :NOT_PERSISTENT
- :INCONCLUSIVE: The persistence test is inconclusive; this function currently cannot determine whether this network is persistent or not.

CatalystNetworkAnalysis.minimalsiphons — Function

minimalsiphons(rs::ReactionSystem)

Constructs the set of minimal siphons of a reaction network, where a siphon is a set of species that can be "switched off," i.e. if the species each have concentration 0, the concentration of all the species will remain 0 for all time. A minimal siphon is one that does not contain a siphon as a strict subset.

The Petri net algorithm relies on detecting siphons in the network, which follows a SAT-solver approach[5] [6].

Concentration Robustness

A species in a reaction network has concentration robustness if its concentration in any steady state of the reaction network is the same. Species with this property will have the same concentration regardless of initial conditions or perturbations to the system.

Deficiency results: these are very limited to deficiency one networks, in which robust species can be detected from the complex composition matrix. Additionally, there is a necessary deficiency-related condition for networks to have ACR, which can be used to detect networks without ACR[3].

Algebraic results: this class of results relies on results from real algebraic geometry. A network is concentration-robust if the locus of steady states lies entirely within a hyperplane where some species concentration is constant. Often, this boils down to finding basis elements of the form $(x_i - \alpha)$ for the steady state ideal[8].

Like persistence, absolute concentration robustness currently cannot be detected conclusively in every case.

CatalystNetworkAnalysis.isconcentrationrobust — Function

isconcentrationrobust(rn::ReactionSystem)

Requires the parameter values be Rational or Integer. Check whether a reaction network has any concentration-robust species. Return codes: 
- :MASS_ACTION_ACR - this species is concentration-robust for the given set of rate constants 
- :GLOBAL_ACR - this species is absolutely concentration-robust for every choice of rate constants
- :INCONCLUSIVE - the algorithm currently cannot decide whether this network has ACR. One could try calling this function with rate constants provided. 
- :NO_ACR - the reaction network does not have ACR. 
- :INEXACTPARAMS - the algorithm cannot conclude concentration-robustness due to inexact parameters (Floats that are too small)

Follows the approach outlined in [Puente et al. 2023](https://arxiv.org/abs/2401.00078).

Misc references[9] [10] [11] [12] [13] [14] [15] [16] [17] [18].

References

1Feinberg, M. Foundations of Chemical Reaction Network Theory; Applied Mathematical Sciences; Springer International Publishing: Cham, 2019; Vol. 202. https://doi.org/10.1007/978-3-030-03858-8.

2Ji, H. Uniqueness of Equilibria for Complex Chemical Reaction Networks. Ph.D. Thesis, The Ohio State University, Columbus, OH, 2011.

3Eloundou-Mbebi, J. M. O.; Küken, A.; Omranian, N.; Kleessen, S.; Neigenfind, J.; Basler, G.; Nikoloski, Z. A Network Property Necessary for Concentration Robustness. Nat Commun 2016, 7 (1), 13255. https://doi.org/10.1038/ncomms13255.

4Angeli, D.; De Leenheer, P.; Sontag, E. D. A Petri Net Approach to the Study of Persistence in Chemical Reaction Networks. Math Biosci 2007, 210 (2), 598–618. https://doi.org/10.1016/j.mbs.2007.07.003.

5Shiu, A.; Sturmfels, B. Siphons in Chemical Reaction Networks. Bull. Math. Biol. 2010, 72 (6), 1448–1463. https://doi.org/10.1007/s11538-010-9502-y.

6Nabli, F.; Martinez, T.; Fages, F.; Soliman, S. On Enumerating Minimal Siphons in Petri Nets Using CLP and SAT Solvers: Theoretical and Practical Complexity. Constraints 2016, 21 (2), 251–276. https://doi.org/10.1007/s10601-015-9190-1.

8Puente, L. D. G.; Gross, E.; Harrington, H. A.; Johnston, M.; Meshkat, N.; Millán, M. P.; Shiu, A. Absolute Concentration Robustness: Algebra and Geometry. arXiv December 29, 2023. http://arxiv.org/abs/2401.00078 (accessed 2024-09-27).

9Johnston, M. D.; Müller, S.; Pantea, C. A Deficiency-Based Approach to Parametrizing Positive Equilibria of Biochemical Reaction Systems. arXiv May 23, 2018. http://arxiv.org/abs/1805.09295 (accessed 2024-09-09).

10Craciun, G.; Jin, J.; Yu, P. Y. An Efficient Characterization of Complex-Balanced, Detailed-Balanced, and Weakly Reversible Systems. SIAM J. Appl. Math. 2020, 80 (1), 183–205. https://doi.org/10.1137/19M1244494.

11Loman, T. E.; Ma, Y.; Ilin, V.; Gowda, S.; Korsbo, N.; Yewale, N.; Rackauckas, C.; Isaacson, S. A. Catalyst: Fast and Flexible Modeling of Reaction Networks.

12Angeli, D.; Ferrell, J. E.; Sontag, E. D. Detection of Multistability, Bifurcations, and Hysteresis in a Large Class of Biological Positive-Feedback Systems. Proceedings of the National Academy of Sciences 2004, 101 (7), 1822–1827. https://doi.org/10.1073/pnas.0308265100.

13Müller, S.; Regensburger, G. Generalized Mass Action Systems: Complex Balancing Equilibria and Sign Vectors of the Stoichiometric and Kinetic-Order Subspaces. SIAM J. Appl. Math. 2012, 72 (6), 1926–1947. https://doi.org/10.1137/110847056.

14Conradi, C.; Feliu, E.; Mincheva, M.; Wiuf, C. Identifying Parameter Regions for Multistationarity. PLoS Comput Biol 2017, 13 (10), e1005751. https://doi.org/10.1371/journal.pcbi.1005751.

15Craciun, G.; Nazarov, F.; Pantea, C. Persistence and Permanence of Mass-Action and Power-Law Dynamical Systems. SIAM J. Appl. Math. 2013, 73 (1), 305–329. https://doi.org/10.1137/100812355.

16Knight, D.; Shinar, G.; Feinberg, M. Sharper Graph-Theoretical Conditions for the Stabilization of Complex Reaction Networks. Math Biosci 2015, 262, 10–27. https://doi.org/10.1016/j.mbs.2015.01.002.

17Müller, S.; Feliu, E.; Regensburger, G.; Conradi, C.; Shiu, A.; Dickenstein, A. Sign Conditions for Injectivity of Generalized Polynomial Maps with Applications to Chemical Reaction Networks and Real Algebraic Geometry. arXiv October 30, 2014. http://arxiv.org/abs/1311.5493 (accessed 2024-08-15).

18Johnston, M. D. Translated Chemical Reaction Networks. Bull Math Biol 2014, 76 (5), 1081–1116. https://doi.org/10.1007/s11538-014-9947-5.